backend/requirements.txt
tensorflow
fastapi
uvicorn
numpy
pillow
python-multipart
requests